Rename menu directory name
